[ALPS09944033] tmem: fix coverity
[Description]
Fix Coverity CERT POS54-C and INCOMPATIBLE_CAST.

[Test]
MT6993 boot passed.

MTK-Commit-Id: 1fa0f7ca823c04f6430da9df2c45dd625c2a6962

CR-Id: ALPS09944033
Feature: trusted_mem

+++ drivers/dma-buf/heaps/mtk_sec_heap_priv.h
@@ -61,7 +61,7 @@ struct mtk_sec_heap_buffer {
 	struct dma_heap *heap;
 	struct list_head attachments;
 	struct mutex lock;
-	unsigned long len;
+	u32 len;
 	struct sg_table sg_table;
 	int vmap_cnt;
 	void *vaddr;
